+++ src/documentation/examples/Icon/Carousel.jsx
@@ -1,0 +1,10 @@
+
+import React from 'react';
+import CarouselIcon from '@bufferapp/ui/Icon/Icons/Carousel';
+
+/** Carousel */
+export default function CarouselIconExample() {
+  return (
+    <CarouselIcon size="large" />
+  );
+}